插入基本ok（差订单插入）

+++ src/main/java/scc1/java_project/deliver_server/dbEntity/Bill.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 public class Bill {
     private Long orderId;
     private Long customerId;
-    private Long charge;
-    private Long batchNumber;
+    private Integer charge;
+    private Integer batchNumber;
     private String payType;
     private Boolean isPaid;
